Make template from admin page

+++ src/main/java/com/aleksandrbogomolov/vote_restaurant/controllers/restaurant/VoteRepositoryController.java
@@ -1,25 +1,24 @@
 package com.aleksandrbogomolov.vote_restaurant.controllers.restaurant;
 
+import com.aleksandrbogomolov.vote_restaurant.model.user.Role;
 import com.aleksandrbogomolov.vote_restaurant.model.user.Vote;
 import com.aleksandrbogomolov.vote_restaurant.service.restaurant.VoteService;
 import com.aleksandrbogomolov.vote_restaurant.util.exception.NotFoundException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
+@Slf4j
 @RequestMapping(value = "vote")
 public class VoteRepositoryController {
-
-    private static Logger logger = LoggerFactory.getLogger(VoteRepositoryController.class);
 
     @Autowired
     private VoteService service;
 
-    @RequestMapping(value = "add", params = {"restaurant"})
+    @RequestMapping(value = "add")
     public String newVote(@RequestParam(name = "restaurant") int restaurantId) {
         Vote vote = null;
         try {
@@ -27,7 +26,7 @@ public class VoteRepositoryController {
         } catch (NotFoundException e) {
             e.printStackTrace();
         }
-        if (vote != null) {
+        if (vote != null && vote.getUser().getRole() == Role.USER) {
             service.delete(100001);
             if (vote.getRestaurant().getId() != restaurantId) {
                 service.save(new Vote(), 100001, restaurantId);
